@@ -2,12 +2,12 @@ require 'spec_helper'
 require 'dpl/cli'
 
 describe DPL::CLI do
-  describe :options do
-    example { described_class.new.options[:app]                           .should be == File.basename(Dir.pwd) }
-    example { described_class.new(:app => 'foo')            .options[:app].should be == 'foo'                  }
-    example { described_class.new("--app=foo")              .options[:app].should be == 'foo'                  }
-    example { described_class.new("--app")                  .options[:app].should be == true                   }
-    example { described_class.new("--app=foo", "--app=bar") .options[:app].should be == ['foo', 'bar']         }
+  describe "#options" do
+    example { expect(described_class.new.options[:app])                           .to eq(File.basename(Dir.pwd)) }
+    example { expect(described_class.new(:app => 'foo')            .options[:app]).to eq('foo')                  }
+    example { expect(described_class.new("--app=foo")              .options[:app]).to eq('foo')                  }
+    example { expect(described_class.new("--app")                  .options[:app]).to eq(true)                   }
+    example { expect(described_class.new("--app=foo", "--app=bar") .options[:app]).to eq(['foo', 'bar'])         }
 
     example do
       described_class.new("--foo.bar=foo", "--foo.baz=012345abcd").options[:foo].
@@ -20,22 +20,22 @@ describe DPL::CLI do
     end
 
     example "error handling" do
-      $stderr.should_receive(:puts).with('invalid option "app"')
+      expect($stderr).to receive(:puts).with('invalid option "app"')
       expect { described_class.new("app") }.to raise_error(SystemExit)
     end
   end
 
-  describe :run do
+  describe "#run" do
     example "triggers deploy" do
       provider = double('provider')
-      DPL::Provider.should_receive(:new).and_return(provider)
-      provider.should_receive(:deploy)
+      expect(DPL::Provider).to receive(:new).and_return(provider)
+      expect(provider).to receive(:deploy)
 
       described_class.run("--provider=foo")
     end
 
     example "error handling" do
-      $stderr.should_receive(:puts).with('missing provider')
+      expect($stderr).to receive(:puts).with('missing provider')
       expect { described_class.run }.to raise_error(SystemExit)
     end
 
