@@ -11,65 +11,65 @@ describe DPL::Provider::CloudFiles do
     described_class.new(DummyContext.new, :username => 'username', :api_key => 'api key', :container => 'travis', :region => 'dfw')
   end
 
-  describe :needs_key? do
+  describe "#needs_key?" do
     example do
-      provider.needs_key?.should == false
+      expect(provider.needs_key?).to eq(false)
     end
   end
 
-  describe :api do
+  describe "#api" do
     example do
-      Fog::Storage.should_receive(:new).with(:provider => 'Rackspace', :rackspace_username => 'username', :rackspace_api_key => 'api key', :rackspace_region => 'dfw')
+      expect(Fog::Storage).to receive(:new).with(:provider => 'Rackspace', :rackspace_username => 'username', :rackspace_api_key => 'api key', :rackspace_region => 'dfw')
 
       provider.api
     end
   end
 
-  describe :check_auth do
+  describe "#check_auth" do
     example do
-      provider.should_receive(:log).with('Authenticated as username')
+      expect(provider).to receive(:log).with('Authenticated as username')
 
       provider.check_auth
     end
   end
 
-  describe :push_app do
+  describe "#push_app" do
     let :files do
       files = double
 
       directory = double
-      directory.stub(:files) { files }
+      allow(directory).to receive(:files) { files }
 
       directories = double
-      directories.should_receive(:get).with('travis').and_return(directory)
+      expect(directories).to receive(:get).with('travis').and_return(directory)
 
       service = double(:service)
-      service.stub(:directories) { directories }
-      provider.stub(:api) { service }
+      allow(service).to receive(:directories) { directories }
+      allow(provider).to receive(:api) { service }
 
       files
     end
 
     example do
-      files.should_receive(:create).with(:key => 'a', :body => 'a body')
-      files.should_receive(:create).with(:key => 'b', :body => 'b body')
-      files.should_receive(:create).with(:key => 'c', :body => 'c body')
+      expect(files).to receive(:create).with(:key => 'a', :body => 'a body')
+      expect(files).to receive(:create).with(:key => 'b', :body => 'b body')
+      expect(files).to receive(:create).with(:key => 'c', :body => 'c body')
 
-      Dir.should_receive(:glob).with('**/*').and_return(['a', 'b', 'c'])
-      File.stub(:open) { |name| "#{name} body" }
+      expect(Dir).to receive(:glob).with('**/*').and_return(['a', 'b', 'c'])
+      allow(File).to receive(:open) { |name| "#{name} body" }
 
       provider.push_app
     end
   end
 
-  describe :deploy do
+  describe "#deploy" do
     example 'Not Found' do
       directories = double
-      directories.stub(:get) { nil }
+      allow(directories).to receive(:get) { nil }
 
       service = double(:service)
-      service.stub(:directories) { directories }
-      provider.stub(:api) { service }
+      allow(service).to receive(:directories) { directories }
+      allow(provider).to receive(:api) { service }
 
       expect { provider.deploy }.to raise_error(DPL::Error, 'The specified container does not exist.')
     end
