@@ -6,36 +6,36 @@ describe DPL::Provider::DotCloud do
     described_class.new(DummyContext.new, :app => 'example', :api_key => 'foo')
   end
 
-  describe :check_auth do
+  describe "#check_auth" do
     example do
-      provider.context.should_receive(:shell).with("echo foo | dotcloud setup --api-key")
+      expect(provider.context).to receive(:shell).with("echo foo | dotcloud setup --api-key")
       provider.check_auth
     end
   end
 
-  describe :check_app do
+  describe "#check_app" do
     example do
-      provider.context.should_receive(:shell).with("dotcloud connect example")
+      expect(provider.context).to receive(:shell).with("dotcloud connect example")
       provider.check_app
     end
   end
 
-  describe :needs_key? do
+  describe "#needs_key?" do
     example do
-      provider.needs_key?.should == false
+      expect(provider.needs_key?).to eq(false)
     end
   end
 
-  describe :push_app do
+  describe "#push_app" do
     example do
-      provider.context.should_receive(:shell).with("dotcloud push example")
+      expect(provider.context).to receive(:shell).with("dotcloud push example")
       provider.push_app
     end
   end
 
-  describe :run do
+  describe "#run" do
     example do
-      provider.context.should_receive(:shell).with("dotcloud -A example www test")
+      expect(provider.context).to receive(:shell).with("dotcloud -A example www test")
       provider.run("test")
     end
   end
