@@ -7,9 +7,9 @@ describe DPL::Provider::Heroku do
     described_class.new(DummyContext.new, :app => 'example', :key_name => 'key', :api_key => "foo", :strategy => "gitdeploykey")
   end
 
-  describe :ssh do
+  describe "#ssh" do
     it "doesn't require an ssh key" do
-      provider.needs_key?.should == false
+      expect(provider.needs_key?).to eq(false)
     end
   end
 end
