@@ -7,18 +7,18 @@ describe DPL::Provider::Heroku do
     described_class.new(DummyContext.new, :app => 'example', :key_name => 'key', :api_key => "foo", :strategy => "git")
   end
 
-  describe :api do
+  describe "#api" do
     it 'accepts an api key' do
       api = double(:api)
-      ::Heroku::API.should_receive(:new).with(:api_key => "foo").and_return(api)
-      provider.api.should be == api
+      expect(::Heroku::API).to receive(:new).with(:api_key => "foo").and_return(api)
+      expect(provider.api).to eq(api)
     end
 
     it 'accepts a user and a password' do
       api = double(:api)
       provider.options.update(:user => "foo", :password => "bar")
-      ::Heroku::API.should_receive(:new).with(:user => "foo", :password => "bar").and_return(api)
-      provider.api.should be == api
+      expect(::Heroku::API).to receive(:new).with(:user => "foo", :password => "bar").and_return(api)
+      expect(provider.api).to eq(api)
     end
   end
 
@@ -30,74 +30,74 @@ describe DPL::Provider::Heroku do
     end
 
     before do
-      ::Heroku::API.should_receive(:new).and_return(api)
+      expect(::Heroku::API).to receive(:new).and_return(api)
       provider.api
     end
 
     its(:api) { should be == api }
 
-    describe :check_auth do
+    describe "#check_auth" do
       example do
-        provider.should_receive(:log).with("authenticated as foo@bar.com")
+        expect(provider).to receive(:log).with("authenticated as foo@bar.com")
         provider.check_auth
       end
     end
 
-    describe :check_app do
+    describe "#check_app" do
       example do
-        provider.should_receive(:log).at_least(1).times.with(/example/)
+        expect(provider).to receive(:log).at_least(1).times.with(/example/)
         provider.check_app
-        provider.options[:git].should be == "GIT URL"
+        expect(provider.options[:git]).to eq("GIT URL")
       end
     end
 
-    describe :setup_key do
+    describe "#setup_key" do
       example do
-        File.should_receive(:read).with("the file").and_return("foo")
-        api.should_receive(:post_key).with("foo")
+        expect(File).to receive(:read).with("the file").and_return("foo")
+        expect(api).to receive(:post_key).with("foo")
         provider.setup_key("the file")
       end
     end
 
-    describe :remove_key do
+    describe "#remove_key" do
       example do
-        api.should_receive(:delete_key).with("key")
+        expect(api).to receive(:delete_key).with("key")
         provider.remove_key
       end
     end
 
-    describe :push_app do
+    describe "#push_app" do
       example do
         provider.options[:git] = "git://something"
-        provider.context.should_receive(:shell).with("git push git://something HEAD:refs/heads/master -f")
+        expect(provider.context).to receive(:shell).with("git push git://something HEAD:refs/heads/master -f")
         provider.push_app
       end
     end
 
-    describe :run do
+    describe "#run" do
       example do
         data = double("data", :body => { "rendezvous_url" => "rendezvous url" })
-        api.should_receive(:post_ps).with("example", "that command", :attach => true).and_return(data)
-        Rendezvous.should_receive(:start).with(:url => "rendezvous url")
+        expect(api).to receive(:post_ps).with("example", "that command", :attach => true).and_return(data)
+        expect(Rendezvous).to receive(:start).with(:url => "rendezvous url")
         provider.run("that command")
       end
     end
 
-    describe :restart do
+    describe "#restart" do
       example do
-        api.should_receive(:post_ps_restart).with("example")
+        expect(api).to receive(:post_ps_restart).with("example")
         provider.restart
       end
     end
 
-    describe :deploy do
+    describe "#deploy" do
       example "not found error" do
-        provider.should_receive(:api) { raise ::Heroku::API::Errors::NotFound.new("the message", nil) }.at_least(:once)
+        expect(provider).to receive(:api) { raise ::Heroku::API::Errors::NotFound.new("the message", nil) }.at_least(:once)
         expect { provider.deploy }.to raise_error(DPL::Error, 'the message (wrong app "example"?)')
       end
 
       example "unauthorized error" do
-        provider.should_receive(:api) { raise ::Heroku::API::Errors::Unauthorized.new("the message", nil) }.at_least(:once)
+        expect(provider).to receive(:api) { raise ::Heroku::API::Errors::Unauthorized.new("the message", nil) }.at_least(:once)
         expect { provider.deploy }.to raise_error(DPL::Error, 'the message (wrong API key?)')
       end
     end
