@@ -7,12 +7,12 @@ describe DPL::Provider::Openshift do
     described_class.new(DummyContext.new, :user => 'foo', :password => 'foo', :domain => 'foo', :app => 'example', :key_name => 'key')
   end
 
-  describe :api do
+  describe "#api" do
     it 'accepts a user and a password' do
       api = double(:api)
       provider.options.update(:user => "foo", :password => "bar")
-      ::RHC::Rest::Client.should_receive(:new).with(:user => "foo", :password => "bar", :server => "openshift.redhat.com").and_return(api)
-      provider.api.should be == api
+      expect(::RHC::Rest::Client).to receive(:new).with(:user => "foo", :password => "bar", :server => "openshift.redhat.com").and_return(api)
+      expect(provider.api).to eq(api)
     end
   end
 
@@ -29,60 +29,61 @@ describe DPL::Provider::Openshift do
     end
 
     before do
-      ::RHC::Rest::Client.should_receive(:new).at_most(:once).and_return(api)
+      expect(::RHC::Rest::Client).to receive(:new).at_most(:once).and_return(api)
       provider.api
     end
 
     its(:api) {should be == api}
 
-    describe :check_auth do
+    describe "#check_auth" do
       example do
-        provider.should_receive(:log).with("authenticated as foo@bar.com")
+        expect(provider).to receive(:log).with("authenticated as foo@bar.com")
         provider.check_auth
       end
     end
 
-    describe :check_app do
+    describe "#check_app" do
       example do
-        provider.should_receive(:log).with("found app example")
+        expect(provider).to receive(:log).with("found app example")
         provider.check_app
       end
     end
 
-    describe :setup_key do
+    describe "#setup_key" do
       example do
-        File.should_receive(:read).with("the file").and_return("ssh-rsa\nfoo")
-        api.should_receive(:add_key).with("key", "foo", "ssh-rsa")
+        expect(File).to receive(:read).with("the file").and_return("ssh-rsa\nfoo")
+        expect(api).to receive(:add_key).with("key", "foo", "ssh-rsa")
         provider.setup_key("the file")
       end
     end
 
-    describe :remove_key do
+    describe "#remove_key" do
       example do
-        api.should_receive(:delete_key).with("key")
+        expect(api).to receive(:delete_key).with("key")
         provider.remove_key
       end
     end
 
-    describe :push_app do
+    describe "#push_app" do
       example "when app.deployment_branch is not set" do
-        provider.app.should_receive :deployment_branch
-        provider.context.should_receive(:shell).with("git push git://something -f")
-        provider.push_app
-      end
-
-      example "when app.deployment_branch is set" do
-        provider.app.stub(:deployment_branch) { "test-branch" }
-
-        provider.context.should_receive(:shell).with("rhc app configure example --deployment-branch test-branch")
-        provider.context.should_receive(:shell).with("git push git://something -f test-branch")
+        expect(provider.context).to receive(:shell).with("git push git://something -f")
         provider.push_app
       end
     end
 
-    describe :restart do
+    context "when app.deployment_branch is set" do
+      subject :provider do
+        described_class.new(DummyContext.new, :user => 'foo', :password => 'foo', :domain => 'foo', :app => 'example', :key_name => 'key', :deployment_branch => 'test-branch')
+
+      expect(provider.app).to receive(:deployment_branch=).with("test-branch")
+      expect(provider.context).to receive(:shell).with("git push git://something -f test-branch")
+      provider.push_app
+      end
+    end
+
+    describe "#restart" do
       example do
-        provider.app.should_receive(:restart)
+        expect(provider.app).to receive(:restart)
         provider.restart
       end
     end
