@@ -7,179 +7,154 @@ describe DPL::Provider do
 
   before { stub_const "DPL::Provider::Example", example_provider }
 
-  describe :new do
-    example { described_class.new(DummyContext.new, :provider => "example") .should be_an(example_provider) }
-    example { described_class.new(DummyContext.new, :provider => "Example") .should be_an(example_provider) }
-    example { described_class.new(DummyContext.new, :provider => "exa_mple").should be_an(example_provider) }
-    example { described_class.new(DummyContext.new, :provider => "exa-mple").should be_an(example_provider) }
+  describe "#new" do
+    example { expect(described_class.new(DummyContext.new, :provider => "example")) .to be_an(example_provider) }
+    example { expect(described_class.new(DummyContext.new, :provider => "Example")) .to be_an(example_provider) }
+    example { expect(described_class.new(DummyContext.new, :provider => "exa_mple")).to be_an(example_provider) }
+    example { expect(described_class.new(DummyContext.new, :provider => "exa-mple")).to be_an(example_provider) }
   end
 
-  describe :requires do
+  describe "#requires" do
     before do
-      example_provider.should_receive(:require).with("foo")
+      expect(example_provider).to receive(:require).with("foo")
     end
 
     example "installed" do
-      example_provider.should_receive(:gem).with("foo", "~> 1.4")
+      expect(example_provider).to receive(:gem).with("foo", "~> 1.4")
       example_provider.requires("foo", :version => "~> 1.4")
     end
 
     example "missing" do
-      example_provider.should_receive(:gem).with("foo", "~> 1.4").and_raise(LoadError)
-      example_provider.context.should_receive(:shell).with('gem install foo -v "~> 1.4"', retry: true)
+      expect(example_provider).to receive(:gem).with("foo", "~> 1.4").and_raise(LoadError)
+      expect(example_provider.context).to receive(:shell).with('gem install foo -v "~> 1.4"', retry: true)
       example_provider.requires("foo", :version => "~> 1.4")
     end
   end
 
-  describe :apt_get do
+  describe "#apt_get" do
     example "installed" do
-      example_provider.should_receive(:`).with("which foo").and_return("/bin/foo\n")
-      example_provider.should_not_receive(:system)
+      expect(example_provider).to receive(:`).with("which foo").and_return("/bin/foo\n")
+      expect(example_provider).not_to receive(:system)
       example_provider.apt_get("foo")
     end
 
     example "missing" do
-      example_provider.should_receive(:`).with("which foo").and_return("")
-      example_provider.context.should_receive(:shell).with("sudo apt-get -qq install foo", retry: true)
+      expect(example_provider).to receive(:`).with("which foo").and_return("")
+      expect(example_provider.context).to receive(:shell).with("sudo apt-get -qq install foo", retry: true)
       example_provider.apt_get("foo")
     end
   end
 
-  describe :pip do
+  describe "#pip" do
     example "installed" do
-      example_provider.should_receive(:`).with("which foo").and_return("/bin/foo\n")
-      example_provider.should_not_receive(:system)
+      expect(example_provider).to receive(:`).with("which foo").and_return("/bin/foo\n")
+      expect(example_provider).not_to receive(:system)
       example_provider.pip("foo")
     end
 
     example "missing" do
-      example_provider.should_receive(:`).with("which foo").and_return("")
-      example_provider.context.should_receive(:shell).with("sudo pip install foo", retry: true)
+      expect(example_provider).to receive(:`).with("which foo").and_return("")
+      expect(example_provider.context).to receive(:shell).with("sudo pip install foo", retry: true)
       example_provider.pip("foo")
     end
   end
 
-  describe :deploy do
-    context 'without newrelic key' do
-      before do
-        provider.should_receive(:check_auth)
-        provider.should_receive(:check_app)
-        provider.should_receive(:push_app)
-        provider.should_receive(:run).with("foo")
-        provider.should_receive(:run).with("bar")
-      end
-
-      example "needs key" do
-        provider.should_receive(:remove_key)
-        provider.should_receive(:create_key)
-        provider.should_receive(:setup_key)
-        provider.should_receive(:setup_git_ssh)
-        provider.deploy
-      end
-
-      example "does not need key" do
-        provider.stub(:needs_key? => false)
-        provider.deploy
-      end
+  describe "#deploy" do
+    before do
+      expect(provider).to receive(:check_auth)
+      expect(provider).to receive(:check_app)
+      expect(provider).to receive(:push_app)
+      expect(provider).to receive(:run).with("foo")
+      expect(provider).to receive(:run).with("bar")
     end
 
-    context 'with newrelic key' do
-      let(:provider) {
-        example_provider.new(
-          DummyContext.new, :app => 'example', :key_name => 'foo', :run => ["foo", "bar"],
-          :notify => [:new_relic => {:api_key => 'foo', :app_name => 'bar'}]
-        )
-      }
+    example "needs key" do
+      expect(provider).to receive(:remove_key)
+      expect(provider).to receive(:create_key)
+      expect(provider).to receive(:setup_key)
+      expect(provider).to receive(:setup_git_ssh)
+      provider.deploy
+    end
 
-      before do
-        provider.should_receive(:check_auth)
-        provider.should_receive(:check_app)
-        provider.should_receive(:push_app)
-        provider.should_receive(:run).with("foo")
-        provider.should_receive(:run).with("bar")
-      end
-
-      example "with newrelic opt" do
-        provider.stub(:needs_key? => false)
-        ::DPL::Notifier::NewRelic.any_instance.stub(:notify).and_return(nil)
-        provider.deploy
-      end
+    example "does not need key" do
+      allow(provider).to receive_messages(:needs_key? => false)
+      provider.deploy
     end
   end
 
-  describe :cleanup do
+  describe "#cleanup" do
     example do
-      provider.context.should_receive(:shell).with('mv .dpl ~/dpl')
-      provider.context.should_receive(:shell).with('git stash --all')
-      provider.context.should_receive(:shell).with('mv ~/dpl .dpl')
+      expect(provider.context).to receive(:shell).with('mv .dpl ~/dpl')
+      expect(provider.context).to receive(:shell).with('git stash --all')
+      expect(provider.context).to receive(:shell).with('mv ~/dpl .dpl')
       provider.cleanup
     end
 
     example "skip cleanup" do
-      provider.options.should_receive(:[]).with(:skip_cleanup).and_return("true")
-      provider.context.should_not_receive(:shell)
+      expect(provider.options).to receive(:[]).with(:skip_cleanup).and_return("true")
+      expect(provider.context).not_to receive(:shell)
       provider.cleanup
     end
   end
 
-  describe :uncleanup do
+  describe "#uncleanup" do
     example do
-      provider.context.should_receive(:shell).with('git stash pop')
+      expect(provider.context).to receive(:shell).with('git stash pop')
       provider.uncleanup
     end
 
     example "skip cleanup" do
-      provider.options.should_receive(:[]).with(:skip_cleanup).and_return("true")
-      provider.context.should_not_receive(:shell)
+      expect(provider.options).to receive(:[]).with(:skip_cleanup).and_return("true")
+      expect(provider.context).not_to receive(:shell)
       provider.uncleanup
     end
   end
 
-  describe :create_key do
+  describe "#create_key" do
     example do
-      provider.context.should_receive(:shell).with('ssh-keygen -t rsa -N "" -C foo -f thekey')
+      expect(provider.context).to receive(:shell).with('ssh-keygen -t rsa -N "" -C foo -f thekey')
       provider.create_key('thekey')
     end
   end
 
-  describe :setup_git_ssh do
+  describe "#setup_git_ssh" do
     after { FileUtils.rm ENV.delete('GIT_SSH') }
 
     example do
       provider.setup_git_ssh('foo', 'bar')
-      ENV['GIT_SSH'].should be == File.expand_path('foo')
+      expect(ENV['GIT_SSH']).to eq(File.expand_path('foo'))
     end
   end
 
-  describe :log do
+  describe "#log" do
     example do
-      $stderr.should_receive(:puts).with("foo")
+      expect($stderr).to receive(:puts).with("foo")
       provider.log("foo")
     end
   end
 
-  describe :shell do
+  describe "#shell" do
     example do
-      example_provider.should_receive(:system).with("command")
+      expect(example_provider).to receive(:system).with("command")
       example_provider.shell("command")
     end
   end
 
-  describe :npm_g do
+  describe "#npm_g" do
     example do
-      example_provider.context.should_receive(:shell).with("npm install -g foo", retry: true)
+      expect(example_provider.context).to receive(:shell).with("npm install -g foo", retry: true)
       example_provider.npm_g("foo")
     end
   end
 
-  describe :run do
+  describe "#run" do
     example do
-      provider.should_receive(:error).with("running commands not supported")
+      expect(provider).to receive(:error).with("running commands not supported")
       provider.run "blah"
     end
   end
 
-  describe :error do
+  describe "#error" do
     example do
       expect { provider.error("Foo") }.to raise_error("Foo")
     end
