@@ -1,7 +1,15 @@
 module DPL
   class Provider
     class Openshift < Provider
-      requires 'rhc'
+      # rhc 1.25.3 and later are required for httpclient 2.4.0 and up
+      # See https://github.com/openshift/rhc/pull/600
+      requires 'httpclient', version: '~> 2.4.0'
+      requires 'rhc', version: '~> 1.25.3'
+
+      def initialize(context, options)
+        super
+        @deployment_branch = options[:deployment_branch]
+      end
 
       def api
         @api ||= ::RHC::Rest::Client.new(:user => option(:user), :password => option(:password), :server => 'openshift.redhat.com')
@@ -33,8 +41,9 @@ module DPL
       end
 
       def push_app
-        if app.deployment_branch
-          context.shell "rhc app configure #{app.name} --deployment-branch #{app.deployment_branch}"
+        if @deployment_branch
+          log "deployment_branch detected: #{@deployment_branch}"
+          app.deployment_branch = @deployment_branch
           context.shell "git push #{app.git_url} -f #{app.deployment_branch}"
         else
           context.shell "git push #{app.git_url} -f"
